+++ src/Composite/AutocompleteComposite/AutocompleteComposite.driver.js
@@ -1,0 +1,12 @@
+import inputAreaWithLabelCompositeDriverFactory from '../InputAreaWithLabelComposite/InputAreaWithLabelComposite.driver';
+
+const autocompleteDriverFactory = ({element, wrapper}) => {
+  const input = element.querySelector('input');
+
+  return {
+    ...inputAreaWithLabelCompositeDriverFactory({element, wrapper}),
+    hasAutocomplete: () => !!input
+  };
+};
+
+export default autocompleteDriverFactory;