+++ src/Composite/AutocompleteComposite/AutocompleteComposite.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import {children, optional, once} from '../';
+import Label from '../../Label';
+import Autocomplete from '../../Autocomplete';
+import InputAreaWithLabelComposite from '../InputAreaWithLabelComposite/InputAreaWithLabelComposite';
+
+const AutocompleteComposite = ({...props, children}) => (
+  <InputAreaWithLabelComposite {...props}>
+    {children}
+  </InputAreaWithLabelComposite>
+);
+
+AutocompleteComposite.propTypes = {
+  children: children(optional(Label), once(Autocomplete))
+};
+
+AutocompleteComposite.displayName = 'AutocompleteComposite';
+
+export default AutocompleteComposite;